fix(test): move TaintTestMethodReturnTypeProvider to own file
PSR-1 requires each class to be in its own file.
